Changed webapplication initialize class.
