--- softgroup/frontend/rest-server/src/main/java/com/softgroup/server/rest/config/WebAppConfig.java
@@ -1,16 +1,0 @@
-package com.softgroup.server.rest.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.view.JstlView;
-import org.springframework.web.servlet.view.UrlBasedViewResolver;
-
-@Configuration
-@EnableWebMvc
-@ComponentScan("com.softgroup.server")
-public class WebAppConfig {
-
-
-}
