+++ softgroup/frontend/rest-server/src/main/java/com/softgroup/server/rest/controller/PrivateController.java
@@ -6,17 +6,11 @@ import org.springframework.web.bind.annotation.*;
 
 @Controller
 @RequestMapping(value = "/api")
-public class MainController {
+public class PrivateController {
 
     @RequestMapping(value = "/private",method = RequestMethod.POST,consumes="application/json",produces = "application/json")
     @ResponseBody
     public Response privateController(@RequestParam String data) {
-        return new Response();
-    }
-
-    @RequestMapping(value = "/public",method = RequestMethod.POST,consumes="application/json",produces = "application/json")
-    @ResponseBody
-    public Response publicController(@RequestParam String data) {
         return new Response();
     }
